파싱부 메인에 연결
